+++ helpers.py
@@ -1,0 +1,22 @@
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+import random
+import string
+
+def click_element(driver, locator):
+
+    element = WebDriverWait(driver, 20).until(
+        EC.element_to_be_clickable(locator))
+    driver.execute_script("arguments[0].scrollIntoView();", element)
+    driver.execute_script("arguments[0].click();", element)
+
+def generate_random_string(length=8):
+    letters = string.ascii_letters
+    return ''.join(random.choice(letters) for i in range(length))
+
+def generate_random_email():
+    return f"{generate_random_string(5)}@example.com"
+
+def generate_random_password(length=8):
+    characters = string.ascii_letters + string.digits + string.punctuation
+    return ''.join(random.choice(characters) for i in range(length))
