+++ test/test_constructor.py
@@ -1,16 +1,7 @@
 import time
 from selenium import webdriver
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
 from locators import Locators
-
-
-def click_element(driver, locator):
-
-    element = WebDriverWait(driver, 20).until(
-        EC.element_to_be_clickable(locator))
-    driver.execute_script("arguments[0].scrollIntoView();", element)
-    driver.execute_script("arguments[0].click();", element)
+import helpers
 
 
 def test_bun_section():
@@ -18,9 +9,9 @@ def test_bun_section():
 
     try:
         driver.get("https://stellarburgers.nomoreparties.site/")
-        click_element(driver, Locators.constructor)
+        helpers.click_element(driver, Locators.constructor)
         time.sleep(4)
-        click_element(driver, Locators.bun_section)
+        helpers.click_element(driver, Locators.bun_section)
         time.sleep(4)
         assert "Булки" in driver.page_source
 
@@ -33,9 +24,9 @@ def test_sauce_section():
 
     try:
         driver.get("https://stellarburgers.nomoreparties.site/")
-        click_element(driver, Locators.constructor)
+        helpers.click_element(driver, Locators.constructor)
         time.sleep(4)
-        click_element(driver, Locators.section_sauces)
+        helpers.click_element(driver, Locators.section_sauces)
         time.sleep(4)
         assert "Соусы" in driver.page_source
 
@@ -48,9 +39,9 @@ def test_fillings_section():
 
     try:
         driver.get("https://stellarburgers.nomoreparties.site/")
-        click_element(driver, Locators.constructor)
+        helpers.click_element(driver, Locators.constructor)
         time.sleep(4)
-        click_element(driver, Locators.filling_section)
+        helpers.click_element(driver, Locators.filling_section)
         time.sleep(4)
         assert "Начинки" in driver.page_source
